requirements.txt
lxml>=4.9.3
playwright>=1.37.0
requests>=2.31.0
selenium>=4.11.2
urllib3>=2.0.4
google-api-python-client>=2.97.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python_dotenv>=1.0.0
sqlalchemy>=2.0.20